# perf(rome_rowan): Refactor `last_token` to iterate from the back