## Summary

The current implementation of `last_token` iterates over the whole sub-tree to find the last token. This is problematic because `last_token` is called frequently inside the formatter:

* To determine if there's any trailing comment
* Inside `join_element` to determine the number of lines between two nodes.

This PR adds a `Direction` to `PreorderTraversal` and uses `Direction::Prev` to get the last token. This reduces the complexity from `O(nodes)` to `O(log(nodes))` (or `O(depth)` to be precise). The sole exception is if the last node is a node without any token (either a list or an array hole. It then is necessary to navigate one level up to the previous token. But that's still `O(log(nodes))`.

This significantly improves the formatter's performance:

```
group                                    last-token                             main
-----                                    ----------                             ----
formatter/checker.ts                     1.00    235.3±2.30ms    11.0 MB/sec    2.70    634.9±3.49ms     4.1 MB/sec
formatter/compiler.js                    1.00    139.4±1.57ms     7.5 MB/sec    2.87    399.4±5.35ms     2.6 MB/sec
formatter/d3.min.js                      1.00    118.0±1.71ms     2.2 MB/sec    10.22  1205.4±9.13ms   222.7 KB/sec
formatter/dojo.js                        1.00      8.0±2.34ms     8.5 MB/sec    2.72     21.8±0.21ms     3.1 MB/sec
formatter/ios.d.ts                       1.00    189.6±2.12ms     9.8 MB/sec    1.83    347.8±5.84ms     5.4 MB/sec
formatter/jquery.min.js                  1.00     32.3±0.38ms     2.6 MB/sec    3.19    103.2±0.91ms   819.8 KB/sec
formatter/math.js                        1.00    228.1±3.81ms     2.8 MB/sec    4.22    962.8±3.69ms   688.7 KB/sec
formatter/parser.ts                      1.00      5.3±0.03ms     9.2 MB/sec    2.43     12.8±0.11ms     3.8 MB/sec
formatter/pixi.min.js                    1.00    128.8±1.20ms     3.4 MB/sec    3.68    474.5±2.50ms   947.2 KB/sec
formatter/react-dom.production.min.js    1.00     38.6±0.39ms     3.0 MB/sec    2.30     88.8±0.80ms  1326.8 KB/sec
formatter/react.production.min.js        1.00  1837.7±23.03µs     3.3 MB/sec    1.93      3.5±0.04ms  1775.8 KB/sec
formatter/router.ts                      1.00      3.8±0.04ms    15.6 MB/sec    2.81     10.8±0.07ms     5.6 MB/sec
formatter/tex-chtml-full.js              1.00    288.8±2.60ms     3.2 MB/sec    3.94   1137.9±4.18ms   820.0 KB/sec
formatter/three.min.js                   1.00    153.2±1.83ms     3.8 MB/sec    5.40   827.6±11.36ms   726.4 KB/sec
formatter/typescript.js                  1.00   912.3±15.55ms    10.4 MB/sec    3.30       3.0±0.17s     3.2 MB/sec
formatter/vue.global.prod.js             1.00     50.3±0.33ms     2.4 MB/sec    3.14    158.2±2.28ms   780.1 KB/sec
```


Most time is now spent allocating and de-allocating heap structures, inside the printer (which is expected), or to test if a line has any suppressions.

## Test Plan

`cargo test`